More refactoring. Moving the client into a separate file, making the package hierarchy deeper and more descriptive,
