--- src/main/kotlin/ostelco/GyordeClient.kt
@@ -1,2 +1,0 @@
-package ostelco
-
